incorparting review comments and fixing windows test issues
