+++ src/main/groovy/io/openliberty/tools/gradle/utils/LooseWarApplication.groovy
@@ -37,57 +37,4 @@ public class LooseWarApplication extends LooseApplication {
         }
         config.addDir(sourceDir, "/")
     }
-
-    /**
-     * get web app source directories
-     * @param project
-     * @return
-     */
-    public static List<Path> getWebSourceDirectoriesToMonitor(Project project) {
-        List<Path> retVal = new ArrayList<Path>();
-        Task warTask = project.getTasks().findByName('war')
-        if (warTask != null) {
-            setWarSourceDir(warTask, retVal)
-        } else if (project.configurations.deploy != null) {
-            setWarSourceDirForDeployDependencies(project, retVal)
-        } else {
-            retVal.add("src/main/webapp")
-        }
-        return retVal;
-    }
-    /**
-     * find war deploy dependencies and add source dir
-     * @param project
-     * @param retVal
-     */
-    private static void setWarSourceDirForDeployDependencies(Project project, ArrayList<Path> retVal) {
-        Task warTask
-        HashMap<File, Dependency> completeDeployDeps = DevTaskHelper.getDeployDependencies(project)
-        for (Map.Entry<File, Dependency> entry : completeDeployDeps) {
-            Dependency dependency = entry.getValue();
-            File dependencyFile = entry.getKey();
-
-            if (dependency instanceof ProjectDependency) {
-                Project dependencyProject = dependency.getDependencyProject()
-                String projectType = FilenameUtils.getExtension(dependencyFile.toString())
-                switch (projectType) {
-                    case "war":
-                        warTask = dependencyProject.getTasks().findByName('war')
-                        if (warTask != null) {
-                            setWarSourceDir(warTask, retVal)
-                        }
-                        break;
-                    default:
-                        break;
-                }
-            }
-        }
-    }
-
-    private static void setWarSourceDir(Task warTask, ArrayList<Path> retVal) {
-        War war = (War) warTask.getProject().war
-        if (war.getWebAppDirectory().getAsFile().get() != null) {
-            retVal.add(war.getWebAppDirectory().get().asFile.toPath())
-        }
-    }
 }
